feat: fl-24 refactor structure of db api
